#157 stage 3 (phase 3a): move the ort-free table-rows tail into tf_core
Extract the second half of predict_table_rows — word-to-cell matching
(docling's post-processor via tf_match, or the geometric fallback) and the
span-expanded rows×cols grid assembly — into tf_core::table_rows, shared
with the browser path. The native predict_table_rows keeps its ort head
(resample crop → predict_table_structure) and delegates the tail. The
parity debug dump (fs) stays native-only (#[cfg(feature = "ml")]) so the
shared code compiles for wasm. Pure code-motion; native output unchanged.

Refs #157
